@@ -1,11 +1,12 @@
 {
     "name": "business-gazeta/atol-api",
-    "description": "PHP библиотека для работы с API Atol",
+    "description": "PHP \u0431\u0438\u0431\u043b\u0438\u043e\u0442\u0435\u043a\u0430 \u0434\u043b\u044f \u0440\u0430\u0431\u043e\u0442\u044b \u0441 API Atol",
     "type": "library",
     "license": "MIT",
     "autoload": {
         "psr-4": {
-            "BusinessGazeta\\AtolApi\\": "src/"
+            "BusinessGazeta\\AtolApi\\": "src/",
+            "BusinessGazeta\\AtolApi\\Tests\\": "tests/"
         }
     },
     "authors": [
@@ -15,8 +16,14 @@
         }
     ],
     "require": {
-        "guzzlehttp/guzzle": "*",
+        "guzzlehttp/guzzle": "^7.8",
         "symfony/validator": "^5.4",
-        "symfony/expression-language": "^5.4"
+        "symfony/expression-language": "^5.4",
+        "doctrine/annotations": "^2.0",
+        "symfony/cache": "^6.4"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^10.5",
+        "symfony/dotenv": "^5.4"
     }
 }
