@@ -7,7 +7,6 @@ use JsonSerializable;
 use Symfony\Component\Validator\Constraints as Assert;
 
 
-
 class AdditionalUserProps extends AbstractObject implements JsonSerializable
 {
     #[Assert\Length(
@@ -17,16 +16,18 @@ class AdditionalUserProps extends AbstractObject implements JsonSerializable
     private string $name;
     #[Assert\Length(
         max: 256,
-        maxMessage: 'Платежный адресс может быть больше чем {{ limit }} символов',
+        maxMessage: 'Платежный адрес может быть больше чем {{ limit }} символов',
     )]
     private string $value;
 
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
-        return [
-            'name' => $this->getName(),
-            'value' => $this->getValue()
-        ];
+        return array_filter(
+            [
+                'name' => $this->getName(),
+                'value' => $this->getValue()
+            ]
+        );
     }
 
     /**
