@@ -8,7 +8,6 @@ use JsonSerializable;
 use Symfony\Component\Validator\Constraints as Assert;
 
 
-
 class AgentInfo extends AbstractObject implements JsonSerializable
 {
     private AgentInfoTypeEnum $type;
@@ -19,17 +18,16 @@ class AgentInfo extends AbstractObject implements JsonSerializable
     #[Assert\Valid]
     private ?MoneyTransferOperator $moneyTransferOperator = null;
 
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
-        $params =
+        return array_filter(
             [
-                'type' => $this->getType()->value
-            ];
-        $params = $this->mergeParams($params, $this->getPayingAgent()?->jsonSerialize(), 'paying_agent');
-        $params = $this->mergeParams($params, $this->getReceivePaymentsOperator()?->jsonSerialize(), 'receive_payments_operator');
-        $params = $this->mergeParams($params, $this->getMoneyTransferOperator()?->jsonSerialize(), 'money_transfer_operator');
-
-        return $params;
+                'type' => $this->getType()->value,
+                'paying_agent' => $this->getPayingAgent(),
+                'receive_payments_operator' => $this->getReceivePaymentsOperator(),
+                'money_transfer_operator' => $this->getMoneyTransferOperator()
+            ]
+        );
     }
 
 
@@ -38,68 +36,47 @@ class AgentInfo extends AbstractObject implements JsonSerializable
         $this->type = $type;
     }
 
-    /**
-     * @return AgentInfoTypeEnum
-     */
     public function getType(): AgentInfoTypeEnum
     {
         return $this->type;
     }
 
-    /**
-     * @param AgentInfoTypeEnum $type
-     */
-    public function setType(AgentInfoTypeEnum $type): void
+    public function setType(AgentInfoTypeEnum $type): AgentInfo
     {
         $this->type = $type;
+        return $this;
     }
 
-    /**
-     * @return PayingAgent|null
-     */
     public function getPayingAgent(): ?PayingAgent
     {
         return $this->payingAgent;
     }
 
-    /**
-     * @param PayingAgent|null $payingAgent
-     */
-    public function setPayingAgent(?PayingAgent $payingAgent): void
+    public function setPayingAgent(?PayingAgent $payingAgent): AgentInfo
     {
         $this->payingAgent = $payingAgent;
+        return $this;
     }
 
-    /**
-     * @return ReceivePaymentsOperator|null
-     */
     public function getReceivePaymentsOperator(): ?ReceivePaymentsOperator
     {
         return $this->receivePaymentsOperator;
     }
 
-    /**
-     * @param ReceivePaymentsOperator|null $receivePaymentsOperator
-     */
-    public function setReceivePaymentsOperator(?ReceivePaymentsOperator $receivePaymentsOperator): void
+    public function setReceivePaymentsOperator(?ReceivePaymentsOperator $receivePaymentsOperator): AgentInfo
     {
         $this->receivePaymentsOperator = $receivePaymentsOperator;
+        return $this;
     }
 
-    /**
-     * @return MoneyTransferOperator|null
-     */
     public function getMoneyTransferOperator(): ?MoneyTransferOperator
     {
         return $this->moneyTransferOperator;
     }
 
-    /**
-     * @param MoneyTransferOperator|null $moneyTransferOperator
-     */
-    public function setMoneyTransferOperator(?MoneyTransferOperator $moneyTransferOperator): void
+    public function setMoneyTransferOperator(?MoneyTransferOperator $moneyTransferOperator): AgentInfo
     {
         $this->moneyTransferOperator = $moneyTransferOperator;
+        return $this;
     }
-
 }
