@@ -25,36 +25,13 @@ class Receipt extends AbstractObject implements JsonSerializable
     )]
     #[Assert\Valid]
     /**
-     * @var Item []
+     * @var Item[] $items
      */
     private array $items;
-    #[Assert\Count(
-        min: 1,
-        max: 10,
-        minMessage: 'Ограничение по количеству от 1 до 10',
-        maxMessage: 'Ограничение по количеству от 1 до 10',
-    )]
-    #[Assert\Valid]
     /**
-     * @var Payment []
+     * @var Payment[] $payments
      */
     private array $payments;
-    #[Assert\Count(
-        min: 1,
-        max: 6,
-        minMessage: 'Ограничение по количеству от 1 до 6',
-        maxMessage: 'Ограничение по количеству от 1 до 6',
-    )]
-    #[Assert\Valid]
-    /**
-     * @var Vat []
-     */
-    private ?array $vats = null;
-    #[Assert\Expression(
-        "value === null || this.isCorrectFloat(value, 100000000)",
-        message: 'Максимальное значение цены – 999999999.99, и 2 знака после запятой',
-    )]
-    private float $total;
     #[Assert\Length(
         max: 16,
         maxMessage: 'Максимальная длина строки – 16 символов.',
@@ -75,105 +52,24 @@ class Receipt extends AbstractObject implements JsonSerializable
     )]
     private ?string $deviceNumber = null;
 
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
-        $params =
+        return array_filter(
             [
-                'client' => $this->getClient(),
-                'company' => $this->getCompany(),
-                'total' => $this->getTotal(),
-                'items' => $this->getItems(),
-                'payments' => $this->getPayments(),
-            ];
-        $params = $this->mergeParams($params, $this->getAgentInfo(), 'agent_info');
-        $params = $this->mergeParams($params, $this->getSupplierInfo(), 'supplier_info');
-        $params = $this->mergeParams($params, $this->getVats(), 'vats');
-        $params = $this->mergeParams($params, $this->getAdditionalCheckProps(), 'additional_check_props');
-        $params = $this->mergeParams($params, $this->getCashier(), 'cashier');
-        $params = $this->mergeParams($params, $this->getAdditionalUserProps(), 'additional_user_props');
-        $params = $this->mergeParams($params, $this->getDeviceNumber(), 'device_number');
-
-        return $params;
-    }
-
-    /**
-     * @return Client
-     */
-    public function getClient(): Client
-    {
-        return $this->client;
-    }
-
-    /**
-     * @param Client $client
-     */
-    public function setClient(Client $client): void
-    {
-        $this->client = $client;
-    }
-
-    /**
-     * @return Company
-     */
-    public function getCompany(): Company
-    {
-        return $this->company;
-    }
-
-    /**
-     * @param Company $company
-     */
-    public function setCompany(Company $company): void
-    {
-        $this->company = $company;
-    }
-
-    /**
-     * @return AgentInfo|null
-     */
-    public function getAgentInfo(): ?AgentInfo
-    {
-        return $this->agentInfo;
-    }
-
-    /**
-     * @param AgentInfo|null $agentInfo
-     */
-    public function setAgentInfo(?AgentInfo $agentInfo): void
-    {
-        $this->agentInfo = $agentInfo;
-    }
-
-    /**
-     * @return SupplierInfo|null
-     */
-    public function getSupplierInfo(): ?SupplierInfo
-    {
-        return $this->supplierInfo;
-    }
-
-    /**
-     * @param SupplierInfo|null $supplierInfo
-     */
-    public function setSupplierInfo(?SupplierInfo $supplierInfo): void
-    {
-        $this->supplierInfo = $supplierInfo;
-    }
-
-    /**
-     * @return array
-     */
-    public function getItems(): array
-    {
-        return $this->items;
-    }
-
-    /**
-     * @param array $items
-     */
-    public function setItems(array $items): void
-    {
-        $this->items = $items;
+            'client' => $this->getClient(),
+            'company' => $this->getCompany(),
+            'total' => $this->getTotal(),
+            'items' => $this->getItems(),
+            'payments' => $this->getPayments(),
+            'agent_info' => $this->getAgentInfo(),
+            'supplier_info' => $this->getSupplierInfo(),
+            'vats' => $this->getVats(),
+            'additional_check_props' => $this->getAdditionalCheckProps(),
+            'cashier' => $this->getCashier(),
+            'additional_user_props' => $this->getAdditionalUserProps(),
+            'device_number' => $this->getDeviceNumber()
+        ]
+        );
     }
 
     /**
@@ -184,12 +80,10 @@ class Receipt extends AbstractObject implements JsonSerializable
         return $this->payments;
     }
 
-    /**
-     * @param array $payments
-     */
-    public function setPayments(array $payments): void
+    public function setPayments(array $payments): Receipt
     {
         $this->payments = $payments;
+        return $this;
     }
 
     /**
@@ -197,15 +91,24 @@ class Receipt extends AbstractObject implements JsonSerializable
      */
     public function getVats(): ?array
     {
-        return $this->vats;
-    }
-
-    /**
-     * @param array|null $vats
-     */
-    public function setVats(?array $vats): void
-    {
-        $this->vats = $vats;
+        $vats = [];
+        foreach ($this->items as $item) {
+            $vat = $item->getVat();
+            $type = $vat->getType()->value;
+            $vats[$type]['type'] = $type;
+            $vats[$type]['sum'][] = $vat->getSum();
+        }
+        return array_values(
+            array_map(
+                static function ($item) {
+                    return [
+                        'type' => $item['type'],
+                        'sum' => array_sum($item['sum'])
+                    ];
+                },
+                $vats
+            )
+        );
     }
 
     /**
@@ -213,80 +116,115 @@ class Receipt extends AbstractObject implements JsonSerializable
      */
     public function getTotal(): float
     {
-        return (float)number_format($this->total, 2);
+        return array_sum(
+            array_map(
+                static function (Item $item) {
+                    return $item->getSum();
+                },
+                $this->getItems()
+            )
+        );
     }
 
-    /**
-     * @param float $total
-     */
-    public function setTotal(float $total): void
+    public function getClient(): Client
     {
-        $this->total = $total;
+        return $this->client;
     }
 
-    /**
-     * @return string|null
-     */
+    public function setClient(Client $client): Receipt
+    {
+        $this->client = $client;
+        return $this;
+    }
+
+    public function getCompany(): Company
+    {
+        return $this->company;
+    }
+
+    public function setCompany(Company $company): Receipt
+    {
+        $this->company = $company;
+        return $this;
+    }
+
+    public function getAgentInfo(): ?AgentInfo
+    {
+        return $this->agentInfo;
+    }
+
+    public function setAgentInfo(?AgentInfo $agentInfo): Receipt
+    {
+        $this->agentInfo = $agentInfo;
+        return $this;
+    }
+
+    public function getSupplierInfo(): ?SupplierInfo
+    {
+        return $this->supplierInfo;
+    }
+
+    public function setSupplierInfo(?SupplierInfo $supplierInfo): Receipt
+    {
+        $this->supplierInfo = $supplierInfo;
+        return $this;
+    }
+
+    public function getItems(): array
+    {
+        return $this->items;
+    }
+
+    public function setItems(array $items): Receipt
+    {
+        $this->items = $items;
+        return $this;
+    }
+
     public function getAdditionalCheckProps(): ?string
     {
         return $this->additionalCheckProps;
     }
 
-    /**
-     * @param string|null $additionalCheckProps
-     */
-    public function setAdditionalCheckProps(?string $additionalCheckProps): void
+    public function setAdditionalCheckProps(?string $additionalCheckProps): Receipt
     {
         $this->additionalCheckProps = $additionalCheckProps;
+        return $this;
     }
 
-    /**
-     * @return string|null
-     */
     public function getCashier(): ?string
     {
         return $this->cashier;
     }
 
-    /**
-     * @param string|null $cashier
-     */
-    public function setCashier(?string $cashier): void
+    public function setCashier(?string $cashier): Receipt
     {
         $this->cashier = $cashier;
+        return $this;
     }
 
-    /**
-     * @return AdditionalUserProps|null
-     */
     public function getAdditionalUserProps(): ?AdditionalUserProps
     {
         return $this->additionalUserProps;
     }
 
-    /**
-     * @param AdditionalUserProps|null $additionalUserProps
-     */
-    public function setAdditionalUserProps(?AdditionalUserProps $additionalUserProps): void
+    public function setAdditionalUserProps(?AdditionalUserProps $additionalUserProps): Receipt
     {
         $this->additionalUserProps = $additionalUserProps;
+        return $this;
     }
 
-    /**
-     * @return string|null
-     */
     public function getDeviceNumber(): ?string
     {
         return $this->deviceNumber;
     }
 
-    /**
-     * @param string|null $deviceNumber
-     */
-    public function setDeviceNumber(?string $deviceNumber): void
+    public function setDeviceNumber(?string $deviceNumber): Receipt
     {
         $this->deviceNumber = $deviceNumber;
+        return $this;
     }
+
     public function isCorrectFloat(float $k, int $max, int $decimals = 2): bool
     {
         $parts = explode('.', $k);
