@@ -32,8 +32,12 @@ class ReceivePaymentsOperator extends AbstractObject implements JsonSerializable
     }
 
 
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
-        return $this->mergeParams([], $this->getPhones(), 'phones');
+        return array_filter(
+            [
+                'phones' => $this->getPhones()
+            ]
+        );
     }
 }
