@@ -9,45 +9,36 @@ class SupplierInfoFull extends SupplierInfo
     private ?string $name = null;
     private string $inn;
 
-    /**
-     * @return string|null
-     */
     public function getName(): ?string
     {
         return $this->name;
     }
 
-    /**
-     * @param string|null $name
-     */
-    public function setName(?string $name): void
+    public function setName(?string $name): SupplierInfoFull
     {
         $this->name = $name;
+        return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getInn(): string
     {
         return $this->inn;
     }
 
-    /**
-     * @param string $inn
-     */
-    public function setInn(string $inn): void
+    public function setInn(string $inn): SupplierInfoFull
     {
         $this->inn = $inn;
+        return $this;
     }
 
 
-
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
-        $params = $this->mergeParams([], $this->getPhones(), 'phones');
-        $params = $this->mergeParams($params, $this->getName(), 'name');
-        $params = array_merge($params, ['inn' => $this->getInn()]);
-        return $params;
+        return array_filter([
+                'phones' => $this->getPhones(),
+                'name' => $this->getName(),
+                'inn' => $this->getInn()
+            ]
+        );
     }
 }
