@@ -2,12 +2,15 @@
 
 namespace BusinessGazeta\AtolApi\Object\Sell;
 
+use BusinessGazeta\AtolApi\Enum\Sell\VatTypeEnum;
 use BusinessGazeta\AtolApi\Object\AbstractObject;
 use JsonSerializable;
-use BusinessGazeta\AtolApi\Enum\Sell\VatTypeEnum;
 use Symfony\Component\Validator\Constraints as Assert;
 
 
+/**
+ * @TODO Необходимо сделать расчет суммы налога автоматическим, исходя из налоговой ставки и суммы
+ */
 class Vat extends AbstractObject implements JsonSerializable
 {
     private VatTypeEnum $type;
@@ -39,11 +42,12 @@ class Vat extends AbstractObject implements JsonSerializable
      */
     public function getSum(): ?float
     {
-        return (float)number_format($this->sum, 2);
+        return (float)number_format($this->sum, 2, '.', '');
     }
 
     /**
      * @param float|null $sum
+     * @return $this
      */
     public function setSum(?float $sum): Vat
     {
@@ -52,13 +56,16 @@ class Vat extends AbstractObject implements JsonSerializable
     }
 
 
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
-        $params = $this->mergeParams([], $this->getType()?->value, 'type');
-        $params = $this->mergeParams($params, $this->getSum(), 'sum');
-        return $params;
-
+        return array_filter(
+            [
+                'type' => $this->getType(),
+                'sum' => $this->getSum()
+            ]
+        );
     }
+
     public function isCorrectFloat(float $k, int $max, int $decimals = 2): bool
     {
         $parts = explode('.', $k);
